+++ app/build.gradle
@@ -37,6 +37,9 @@ android {
 
 dependencies {
 
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.3.5'
+    implementation 'androidx.navigation:navigation-ui-ktx:2.3.5'
+
     //서버 연결을 위한 Retrofit2
     implementation "com.squareup.retrofit2:retrofit:2.9.0"
     //Retrofit2에서 gson 사용을 위한 컨버터
