+++ app/src/main/java/org/sopt/androidassignment1/data/local/FollowerData.kt
@@ -1,6 +1,7 @@
-package org.sopt.androidassignment1.follower
+package org.sopt.androidassignment1.data.local
 
 data class FollowerData(
     val name : String,
     val introduction : String
 )
+
