+++ app/src/main/java/org/sopt/androidassignment1/data/local/RepositoryData.kt
@@ -1,6 +1,7 @@
-package org.sopt.androidassignment1.repository
+package org.sopt.androidassignment1.data.local
 
 data class RepositoryData(
     val title : String,
     val content : String
 )
+
