+++ app/src/main/java/org/sopt/androidassignment1/data/remote/RequestSignupData.kt
@@ -1,4 +1,4 @@
-package org.sopt.androidassignment1.signup
+package org.sopt.androidassignment1.data.remote
 
 data class RequestSignupData(
     val name: String,
