+++ app/src/main/java/org/sopt/androidassignment1/data/remote/ResponseSigninData.kt
@@ -1,4 +1,4 @@
-package org.sopt.androidassignment1.signIn
+package org.sopt.androidassignment1.data.remote
 
 data class ResponseSigninData(
     val status : Int,
