+++ app/src/main/java/org/sopt/androidassignment1/data/remote/ResponseSignupData.kt
@@ -1,4 +1,4 @@
-package org.sopt.androidassignment1.signup
+package org.sopt.androidassignment1.data.remote
 
 data class ResponseSignupData(
     val status : Int,
