+++ app/src/main/java/org/sopt/androidassignment1/ui/view/ViewPagerFollowerFragment.kt
@@ -1,4 +1,4 @@
-package org.sopt.androidassignment1.baseFragment
+package org.sopt.androidassignment1.ui.view
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,14 +7,14 @@ import android.view.View
 import android.view.ViewGroup
 import org.sopt.androidassignment1.R
 
-class CameraFragment : Fragment() {
+class ViewPagerFollowerFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_camera, container, false)
+        return inflater.inflate(R.layout.fragment_view_pager_follower, container, false)
     }
 
 }
