+++ app/src/main/java/org/sopt/androidassignment1/util/ViewExt.kt
@@ -1,0 +1,8 @@
+package org.sopt.androidassignment1.util
+
+import android.content.Context
+import android.widget.Toast
+
+fun Context.shortToast(message: String) {
+    Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
+}
